requirements.txt
streamlit==1.19.0
openai==0.27.0
langchain==0.0.208
langchain-community==0.3.21  
python-dotenv==0.21.1  
tiktoken==0.9.0  
